+++ xtcryptosignals/client/api/tools/views.py
@@ -30,6 +30,7 @@ bp = Blueprint("tools", __name__)
 def before_request():
     g.SYMBOLS_PER_EXCHANGE, _ = service.get_symbols_per_exchange()
     g.HISTORY_FREQUENCY, _ = service.get_history_frequency()
+    g.COINS_OR_TOKENS_REFERENCE, _ = service.get_coins_or_tokens_reference()
 
 
 @bp.context_processor
@@ -62,8 +63,8 @@ def coin_or_token_frequency(coin_or_token):
         template_name_or_list="tools/coin_token_data.html",
         socket_base_url=current_app.config["SOCKET_BASE_URL"],
         frequency=g.HISTORY_FREQUENCY[0],
-        # TODO: Make this is setting
-        frequencies=["10s", "1m", "10m", "30m", "1h", "4h", "12h", "1d"],
+        frequencies_charts=["10s", "1m", "10m", "30m", "1h", "4h", "12h", "1d"],
         attributes=["Price USDT"],
         coin_or_token=coin_or_token,
+        reference=g.COINS_OR_TOKENS_REFERENCE[coin_or_token],
     )