+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All changes will be registered here per release.
 ## [0.3.0] - Current date
 Added Crypto Fear & Greed Index in UI.    
 Added new LTO Network Token Binance pairs.  
+Added Charts per Token USD pair.  
 Code fixes and improvements.  
 
 ## [0.2.0] - 2020-12-25